pyproject.toml
[tool.poetry]
name = "project"
version = "1.2.0"
description = "Storasense-Platform"
authors = []
package-mode= false

[tool.poetry.dependencies]
python = "^3.13"
python-semantic-release = "^10.2.0"
uvicorn = "^0.35.0"
fastapi = "^0.116.1"
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
pytest-asyncio = "^1.1.0"
requests-mock = "^1.12.1"
pre-commit = "^4.2.0"
ruff = "^0.12.4"
fastapi-events = "^0.12.2"
alembic = "^1.16.4"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
fastapi-limiter = "^0.1.6"
python-multipart = "^0.0.20"
httpx = "^0.28.1"
coverage = "^7.9.2"
pytest-dotenv = "^0.5.2"
structlog = "^25.4.0"
paho-mqtt = "^2.1.0"
python-dotenv = "^1.1.1"
#pysqlite3 = "^0.5.4"
sqlalchemy = "^2.0.41"
pg8000 = "^1.31.4"
authlib = "^1.6.1"
itsdangerous = "^2.2.0"
urllib3 = "^2.5.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.8.3"


[tool.poetry.group.test.dependencies]
selenium = "^4.35.0"
psycopg2-binary = "^2.9.10"
pytest-dependency = "^0.6.0"
numpy = "^2.3.2"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 79
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = "test"
filterwarnings = [
    "ignore::pydantic._internal._config.PydanticDeprecatedSince20",
    "ignore::sqlalchemy.exc.SAWarning"
]

[tool.semantic_release]
assets = []
build_command_env = []
version_toml = ["pyproject.toml:tool.poetry.version"]
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = false
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
changelog_file = ""
exclude_commit_patterns = []
mode = "update"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = true

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
token = { env = "GITHUB_TOKEN" }
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
